refactor(frontend): extract generatePassword to utils/password.ts
Move generatePassword from user-create.tsx to a dedicated utility file.
Update imports in user-create.tsx, user-edit-modals.tsx, add-user-step.tsx,
and their test mocks.
